Move DepthFirstIteration to org.mastodon.util
